feat: move AuthModal to wallet package
