+++ src/components/index.ts
@@ -44,6 +44,6 @@ export * from './SwapDetails';
 export * from './TagGroup';
 export * from './SpendingLimit';
 export * from './Link';
-export * from './AuthModal';
+export * from '../packages/wallet/AuthModal';
 export * from './MarkdownEditor';
 export * from './MarkdownViewer';
