+++ src/packages/wallet/AuthModal/AccountDetails/styles.ts
