+++ src/packages/wallet/AuthModal/constants.ts
