+++ src/packages/wallet/AuthModal/types.ts
